Refactored POButtonStyle
