--- ui-core/src/main/kotlin/com/processout/sdk/ui/core/style/button/POButtonDefaults.kt
@@ -1,6 +1,0 @@
-package com.processout.sdk.ui.core.style.button
-
-object POButtonDefaults {
-    const val PADDING_HORIZONTAL_DP = 24
-    const val PADDING_VERTICAL_DP = 8
-}
